+++ UI/Services/MonoGameService.cs
@@ -26,6 +26,7 @@
         private readonly ISceneService _sceneService;
         private ComponentEditingStyle _editingStyle = ComponentEditingStyle.Translation;
         private bool _hasSelectedGizmoBeenManuallyChanged = false;
+        private SamplerStateType _samplerStateType = SamplerStateType.PointClamp;
         private bool _showGrid = true;
         private bool _showSelection = true;