+++ CHANGELOG.md
@@ -10,6 +10,8 @@
 * The Layer Collision Matrix can be edited to specify which layers collide with which layers in the physics system.
 * A draggable splash screen shows when the project is loaded on startup.
 * There are now 16 layers instead of 8.
+* The sampler state can be changed on a camera for a sprite batch.
+	* This allows for crisp pixel art.
 
 ## Refactors